requirements.txt
boto3==1.3.1
botocore==1.4.41
docutils==0.12
futures==3.0.5
jmespath==0.9.0
python-dateutil==2.5.3
python-debian==0.1.28
six==1.10.0
